Book your Pelvic Ultrasound Scan at Butler Medical Imaging. (08) 9544 3999
A pelvic ultrasound scan is a key diagnostic tool used to visualize the organs within the pelvis. This non-invasive imaging technique is vital for assessing the health of pelvic structures such as the uterus, ovaries, fallopian tubes, and bladder. Widely used in gynecology and obstetrics, pelvic ultrasounds provide valuable insights into various medical conditions and play a crucial role in early diagnosis, treatment planning, and monitoring of ongoing health issues.